Redefine point array data in serialized YAML format
It had unnecessary values for deserialization and separator was all colons
which makes it hard to read.  Making the point list JavaScript array could
be a way, but it's too redundant.

At last, we have consulted SVG (scalable vector graphics) schema, which
contains point coordinate lists as strings.

This is not backward-compatible modification.