Prohibit uploading a file with name containing path delimiter
This is a security issue.

The ability to specify a path to upload could be a serious security hole
since relative path (../) or even root path (/etc/passwd) could be
specified, and the permission could be misconfigured.

Allowing subdirectories in the data directory could be useful, but for
the time being, it's not worth the risk.