Add a config file for configurable options
The config file is just a php file that defines the "conf" variable.
Since it can change among servers, it is not under version control.
It's automatically generated in the first invocation of canvas.php,
and the user can edit the file for their preferences.

Also added .htaccess to prevent peeking the config file from outside.