Make the menu buttons be a bit object oriented
No functionality is affected by this commit.

We had to maintain the captions and the effect of the menu items
separately in the source.  It's very cumbersome to maintain the items
because an index is the only identity of a menu item.
It's paticularly annoying when an item is inserted or removed.

Other GUI elements such as toolbar buttons and color palette should be
modified like this way, but it's a good starting point.