Add offset to erase gaps in view mode caused by edit mode's toolbar
The canvas has the toolbar and the menu bar in edit mode, but they
are hidden in view mode.  It means blank areas appear when in view
mode which are not present in edit mode in top and left of the
figure.

We erase these areas by offsetting the figure's coordinates in
edit mode.  It means the top left corner of the editable area has
the coordinates (0,0) and that point will be the top left corner in
view mode.