Add meta element in the serialized figure text to define size
The first element in the YAML formatted object array is the meta
object, which defines width and height of the figure.  If it's not
present, default values are used and explicitly included in the
saved text with this and later versions of SketchCanvas.

The size of current figure can be edited by "Size" button in the
menu bar.  It's the first attempt to show input controls on a
layer on top of the canvas.

Width and height values are used in view mode to trim the canvas
to fit its contents.