Insert scripts into header section
Previously, SketchCanvas scripts are embedded in place of <skcanvas>
tags.  It means there will be multiple script inclusions if there are
multiple figures.  The browsers are smart enough to load the same script
only once, but it's bad style to embed scripts in the HTML body.

As a DokuWiki plugin, JavaScript files should be included in the
unified script stream which will be downloaded from js.php, but we
could not make SketchCanvas.js work with compress option (which
discards whitespace), so we added <script> tags in the header.

The reason SketchCanvas.js does not work is probably something to do with
the fact that it doesn't also pass JSLint.  I'm not sure.