Cleaned up an unused function and checking presence of event handlers
We only checked boolean value of event handlers of SketchCanvas, but
they need to be checked if they're member of the object by 'in'
operator first.

Also, ajaxsave() is no longer used by anyone, thus deleted.

saveDataNew() now requires an argument that specifies the name of
the newly saved figure.