Make the SketchCanvas attached canvas reset its size on assignment
If the canvas is used in edit mode and no data is provided, the
original size of the canvas element is retained, but the original
sie may not be appropriate because editing mode requires various
toolbars to be shown.  We could reduce a figure's size, but it
should not affect the canvas size in edit mode.