Add a text input layer instead of prompt()
When the text tool is selected and somewhere in the canvas is clicked,
a layer to input the text is displayed.  Previously, it was the
browser-dependent dialog box created by prompt().

The function prompt() was not only ugly but also blocks the execution
of JavaScript, which can be critical for user experience.