First steps toward objectism in SketchCanvas.js
Draw objects are now proper class objects instead of tables.
This way, we could define methods and override them.

The problem with JavaScript classes is that there are multiple ways
to implement inheritance and no standard way is defined.
We use prototype chain for our DrawObject class tree, but do not
expect it will also apply to outside of SketchCanvas.

It doesn't seem to produce much benefit yet, but it'll be more
apparent as types of drawing objects increase.