Add ability to alter an existing text shape by clicking on it
Normally, clicking on empty space in the canvas with the text tool
leads to new text mode.  Clicking on an existing text shape will
open the text editor with replace mode.

Previously, when the user wanted to change a text, he had to delete
it once and add new one with updated content.  Now this manipulation
should be way easier.