Switch PHP YAML parser library from yaml to Spyc
Spyc is pure PHP implementation of YAML parser.  Yaml, on the other hand,
requires libyaml (C implementation) installed on the system and its
corresponding PHP extension, which are both cumbersome to install.

Due to pure PHP implementation nature, Spyc could be slower than libyaml
(who could beat C implementation in speed, anyway?), but our usage
doesn't invoke them often and the data won't be so big nor complex.
We prefer ease of installation.  It's particularly useful when the
PHP extensions or system-wide package management (yum or apt-get)
are not under the user's control.