(Compatibility warning!) Rename plugin to sketchcanvas
The plugin directory must be named "sketchcanvas" instead of "skcanvas"
from now on.  Inconsistent directory name will make the plugin not
working properly.

The plugin sources have some hard-coded paths to the plugin's directory,
and DokuWiki plugin loader recognizes class names with a pattern that
includes directory name.

This backward-incompatible change has done because the plugin is listed
in the DokuWiki's badextensions page:

https://www.dokuwiki.org/devel:badextensions

for an accusation of plugin name.