Add Noto Sans CJK font for consistent typeface in the server
The font file is large, but putting the file on the server is the
only reliable way of making server-side rendered font independent
of the server's environment.

We do not actually add the font file "NotoSansCJSjp-Regular.otf",
since it's not source and not everyone need server-side renderer or
PDF exporting.

In order to render texts in the sketches with the server-side
renderer, you have to download the font from:

https://noto-website.storage.googleapis.com/pkgs/NotoSansCJKJP-hinted.zip

which is also mentioned in README.md.