Use Noto Sans Japanese font in the canvas, too
Previously, the browsers rendered text with arbitrary font, which
means the text can look different from server-side rendered sketches
or even other browsers.

Using a specific font resolves this issue, but the HTML source must
define @font-face with corresponding font url.  Actually, this is
done by:

@import url(http://fonts.googleapis.com/earlyaccess/notosansjapanese.css);

The @import is preferable because it has several font file formats to
fall back if the browser does not support one of them.