Change plugin base name to sketchcanvas and update date
The plugin base name is the directory's name and it's assumed to
be the same as the DokuWiki official site's page name.  Since we
already have a plugin page named sketchcanvas (which can be reached
from the URL below), we cannot change the page name.

https://www.dokuwiki.org/plugin:sketchcanvas

So we have to change the base name in plugin.info.txt.

Also, it's more unlikely to have a name collision if we have a longer
name.