Support arc, arcarrow and arcbarrow shapes in server-side renderer
PHP's GD library doesn't support Bezier spline curves directly, so
we had to implement them by ourselves.

Actually, it was not that hard.
Reading the Wikipedia alone is enough to implement it by oneself.
https://en.wikipedia.org/wiki/B%C3%A9zier_curve

It could be slow because the curve is actually a collection of lines
after all, but we don't want to hack into GD library itself right now.