Enable live preview of shapes being added
Previously, we could not see the shape being added until we finish
clicking on the final point.  This behavior had two problems:

* We could not tell if we're currently in progress of adding a shape.
* It's hard to lay out curves without seeing any hint.

So we made the shape appear before its final point is determined.
Rendering shapes everytime the mouse moves on the canvas consumes
computation time, but it's already done in size changing by the
bounding box, so we don't care.

In this modification, the rendering procedure is significantly
changed.  SketchCanvas object no longer has "arr" and "idx"
almost-global variables, which were used for managing point
coordinates for the new shape.  Instead, a temporary object named
"cur_shape" is added for storing the shape being added.

Also fixed TextShape's bounding box calculation was wrong for
width == 2.