Add path shape that can have arbitrary number of vertices
It can be used as a polyline, but you can add Bezier spline control
points to that path to make it a curve.

Pressing the Ctrl key while start dragging a vertex would pull out
a control point handle, which can be used to manipulate the curve's
shape.

There was a difficult decision whether the Bezier control points
should be points in the vertex list or just attributes of the
vertices.  For example, former proposal would format a cubic Bezier
spline as 'x0,y0:x1,y1:x2,y2:x3,y3', which looks similar to
existing shape's point list specification, but you cannot tell
whether the coordinates represent an actual vertex position or a
control point.  In the latter case, the format would be like
'x0,y0:cx1,cy1,dx1,dy1,x1,y1' which would make more sense and
naturally extend existing shape formats.

We picked the latter and added "cx", "cy", "dx", "dy" properties
to the vertex.  This change is not incorporated into serialization,
so the path shapes cannot be serialized correctly yet.

There is an issue that the bounding box of a curved path is not
correct when selected.  The bounding box also affects hit detection with
the mouse cursor.