Add option to draw arrowheads at the head and the tail of a path
There are now 3 tools for paths: path without arrowheads, path
with an arrowhead at one end, path with arrowheads at both ends.
This is similar to the relationships between arc, arcarrow and
arcbarrow tools, but their serialized YAML representation is
based on a different concept.

The path object has a new property "arrow" which is a set of
{head, tail}.  Actually, it's serialized as a sequence like
[head, tail] in YAML data, but the order is insignificant.
This format was necessary because JavaScript doesn't natively
support set type.

Traditionally, shapes such as lines and arcs have distinct type
identifiers for variants with arrowheads, like arcarrow or arcbarrow,
but we didn't follow that convention in "path" shape.

So, the 3 tools for the paths are all "path" tool, except that their
initial value for arrow parameter are different.

This is a tough decision but having a dedicated property for
indicating arrowheads presense for each end is preferable in the
case of path, because the path can be so complex that modifying
a property could be far easier than recreating the whole path.
Lines and arcs are easy to recreate from scratch, so having the
arrow property barely helps.

We haven't implemented a way to modify existing path's arrow
property with GUI, so modifying the source YAML text is the only
way to change existing path's arrow property.