Refine behavior of select tool on overlapped shapes
Previously, the select tool first changes selection to the topmost
shape before sizing or moving.  This is really annoying when
manipulating overlapped shapes, especially when a smaller shape
is under a larger one.

Now the select tool respects currently selected shapes and do not
reset the selection if the mouse position is included by the
selected shapes' bounding boxes.