Make pathedit tool's delete key just delete single vertex
All other tools delete all selected shapes, but the user won't expect
pathedit tool to work that way.

Now, the toolset can create a path, move a vertex, modify control
points of a vertex and delete a vertex.  The only missing feature
for being a full-fledged vector graphics path editor is to insert a
vertex to existing path.  This feature is difficult in a different
level than the features we introduced so far, so we would stop here
for the time being.