Prevent an error on https font link
Chrome, Microsoft Edge and IE11 (and probably others too) complains about
notosansjapanese.css is not loaded over https if the page itself is
requested via https.  This is because the URL of file on googleapi server
is hardcoded with http, which bothers those browsers.

Thankfully DokuWiki provides a convenient function to query if the protocol
is https (apparently there are no straightforward way to do this with only
pure PHP APIs), which is is_ssl().  This may be a bit low-tech fix, but we
switch the schema on URL based on the return value of is_ssl().